feat:#54 병원 세부 정보 및 리뷰 섹션의 데이터 로딩 최적화
